# public/docs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Stugether API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/forums": {
            "get": {
                "tags": [
                    "Forums"
                ],
                "summary": "List forums",
                "operationId": "a3d6afb071d6291527c85827e54e11c3",
                "parameters": [
                    {
                        "name": "scope",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "mine",
                                "public",
                                "all"
                            ]
                        }
                    },
                    {
                        "name": "q",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "created_at",
                                "nama"
                            ]
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Forums"
                ],
                "summary": "Create forum",
                "operationId": "fb1118d607a9ed2d476d40a9744a2548",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nama"
                                ],
                                "properties": {
                                    "nama": {
                                        "type": "string"
                                    },
                                    "deskripsi": {
                                        "type": "string"
                                    },
                                    "jenis_forum": {
                                        "type": "string",
                                        "enum": [
                                            "akademik",
                                            "proyek",
                                            "komunitas",
                                            "lainnya"
                                        ]
                                    },
                                    "is_public": {
                                        "type": "integer",
                                        "enum": [
                                            0,
                                            1
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/forums/recommended": {
            "get": {
                "tags": [
                    "Forums"
                ],
                "summary": "Recommended public forums",
                "operationId": "b28e0877b919c8b1102f20b64975fa44",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/forums/{id}": {
            "get": {
                "tags": [
                    "Forums"
                ],
                "summary": "Show forum",
                "operationId": "d5cca19360614be34872c8d1125763be",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Forums"
                ],
                "summary": "Delete forum (admin)",
                "operationId": "d7115cc63ac21168b742e03e8165a253",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Forums"
                ],
                "summary": "Update forum (admin)",
                "operationId": "25b96ffda210e8ca31fd29d79485e09d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nama": {
                                        "type": "string"
                                    },
                                    "deskripsi": {
                                        "type": "string"
                                    },
                                    "jenis_forum": {
                                        "type": "string"
                                    },
                                    "is_public": {
                                        "type": "integer",
                                        "enum": [
                                            0,
                                            1
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}